직원관리 수정 끝! (프로필만 남음) -191214

See merge request whyj97/hagong!174

+++ src/hagong/academy/mngStaff/model/service/StaffService.java
@@ -44,34 +44,39 @@ public class StaffService {
 		int result1 = new StaffDao().updateStaffInfo(con, staff);
 		if(result1 > 0) {
 			commit(con);
-			for(int i = 0; i < fileList.size(); i++) {
-				fileList.get(i).setUserNo(staff.getUserNo());
+			if(fileList != null) {
+				for(int i = 0; i < fileList.size(); i++) {
+					fileList.get(i).setUserNo(staff.getUserNo());
+				}
 			}
 			
 			int result2 =  new StaffDao().updateStaffSubInfo(con, staff);
 			if(result2 > 0) {
 				commit(con);
+				result = 1;
 				
-				int result3 = new StaffDao().insertFile(con, fileList);
-				if(result3 > 0) {
-					commit(con);
-					
-					for(int i = 0; i < fileList.size(); i++) {
-						if(!fileList.get(i).getFileType().equals("PROFILE")) {
-							int fileNo = new StaffDao().selectFileNo(con, fileList.get(i));
-							fileList.get(i).setFileNo(fileNo);
-							int result4 = new StaffDao().insertStaffFile(con, fileList.get(i));
-							if(result4 > 0) {
-								commit(con);
-								result = 1;
-							} else {
-								rollback(con);
-								result = 0;
+				if(fileList != null) {
+					int result3 = new StaffDao().insertFile(con, fileList);
+					if(result3 > 0) {
+						commit(con);
+						
+						for(int i = 0; i < fileList.size(); i++) {
+							if(!fileList.get(i).getFileType().equals("PROFILE")) {
+								int fileNo = new StaffDao().selectFileNo(con, fileList.get(i));
+								fileList.get(i).setFileNo(fileNo);
+								int result4 = new StaffDao().insertStaffFile(con, fileList.get(i));
+								if(result4 > 0) {
+									commit(con);
+									result = 1;
+								} else {
+									rollback(con);
+									result = 0;
+								}
 							}
 						}
+					} else {
+						rollback(con);
 					}
-				} else {
-					rollback(con);
 				}
 			} else {
 				rollback(con);
@@ -80,14 +85,16 @@ public class StaffService {
 			rollback(con);
 		}
 		
-		int delResult = 0;
-		for(int i = 0; i < deleteFile.size(); i++) {
-			delResult = new StaffDao().deleteFile(con, staff.getUserNo(), deleteFile.get(i));
-		}
-		if(delResult > 0) {
-			commit(con);
-		} else {
-			rollback(con);
+		int delResult = 1;
+		if(deleteFile != null) {
+			for(int i = 0; i < deleteFile.size(); i++) {
+				delResult = new StaffDao().deleteFile(con, staff.getUserNo(), deleteFile.get(i));
+			}
+			if(delResult > 0) {
+				commit(con);
+			} else {
+				rollback(con);
+			}
 		}
 		
 		if(result > 0 && delResult > 0) {
